+++ src/reports/orl/sidebar/sidebar.html
@@ -18,7 +18,7 @@
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/intercom.js/0.1.4/intercom.min.js"></script>
 
-    <?!= include('SidebarStylesheet'); ?>
+    <?!= include('reports/orl/sidebar/sidebar.stylesheet'); ?>
   </head>
   <body>
     <div id="app">
@@ -81,6 +81,6 @@
         isSignedIn: "<?= isSignedIn ?>" == "true",
       };
     </script>
-    <?!= include('SidebarScript'); ?>
+    <?!= include('reports/orl/sidebar/sidebar.script'); ?>
   </body>
 </html>
