+++ src/shared/alert.ts
@@ -1,0 +1,3 @@
+export function alert(msg: string) {
+  Browser.msgBox(msg, Browser.Buttons.OK);
+}
