+++ src/shared/include.ts
@@ -1,0 +1,3 @@
+export function include(filename: string) {
+  return HtmlService.createHtmlOutputFromFile(filename).getContent();
+}
