+++ src/shared/round.ts
@@ -1,0 +1,3 @@
+export function round(num: number) {
+  return Math.round((num + Number.EPSILON) * 100) / 100;
+}
